fix(orchestra): relocate DispatchCoordinatorProtocol to KERNEL layer
## Problem

- heartbeat.py imported from vibe3.domain (OBSERVATION layer) - KERNEL
  must not import from OBSERVATION per architecture
- Import from vibe3.domain.protocols.dispatch_protocols was rejected
  as a deep import by modularity tests
- Import from vibe3.domain was rejected as domain business entity
  import by core budget tests
- Import from vibe3.orchestra.domain_types was still a deep import

## Solution

1. Add DispatchCoordinatorProtocol to orchestra/domain_types.py
   (the canonical home for protocols used by KERNEL modules)
2. Export it from orchestra/__init__.py (public API)
3. Import from vibe3.orchestra in heartbeat.py
4. Remove duplicate from domain/protocols/dispatch_protocols.py

## Architecture Compliance

- Follows orchestra/domain_types.py docstring: "KERNEL modules
  must import from this module instead of from vibe3.domain"
- No deep imports (public API only)
- No KERNEL → OBSERVATION dependency direction violation
- Follows modularity-standards.md: public API imports only

## Verification

- All 21 heartbeat tests pass
- All 4 module API compliance tests pass
- All 4 core budget tests pass
- mypy type check passes
