+++ src/vibe3/runtime/heartbeat.py
@@ -9,10 +9,9 @@ from typing import TYPE_CHECKING, Protocol
 
 from loguru import logger
 
-from vibe3.domain import DispatchCoordinatorProtocol
 from vibe3.models import OrchestraConfig
 from vibe3.observability import append_orchestra_event, append_orchestra_run_separator
-from vibe3.orchestra import ServiceBase
+from vibe3.orchestra import DispatchCoordinatorProtocol, ServiceBase
 from vibe3.utils import PACK_REFS_INTERVAL_TICKS
 
 from .periodic_check_executor import execute_periodic_check
